+++ api/routes/users.php
@@ -1,7 +1,7 @@
 <?php
 $app->get('/users', function ($request, $response, $args) {
 	$ib = ($request->getAttribute('ib')) ?? new ib();
-    if ($ib->rbac->checkAccess("ADMIN-USERS")) {
+    if ($ib->auth->checkAccess("ADMIN-USERS")) {
         $ib->api->setAPIResponseData($ib->auth->getAllUsers());
     }
 	$response->getBody()->write(jsonE($GLOBALS['api']));
@@ -13,7 +13,7 @@ $app->get('/users', function ($request, $response, $args) {
 $app->post('/users', function ($request, $response, $args) {
     $ib = ($request->getAttribute('ib')) ?? new ib();
     $data = $ib->api->getAPIRequestData($request);
-    if ($ib->rbac->checkAccess("ADMIN-USERS")) {
+    if ($ib->auth->checkAccess("ADMIN-USERS")) {
         $UN = $data['un'] ?? exit($ib->api->setAPIResponse('Error','Username missing from request'));
         $PW = $data['pw'] ?? exit($ib->api->setAPIResponse('Error','Password missing from request'));
         $FN = $data['fn'] ?? null;
@@ -32,7 +32,7 @@ $app->post('/users', function ($request, $response, $args) {
 $app->patch('/user/{id}', function ($request, $response, $args) {
 	$ib = ($request->getAttribute('ib')) ?? new ib();
     $data = $ib->api->getAPIRequestData($request);
-    if ($ib->rbac->checkAccess("ADMIN-USERS")) {
+    if ($ib->auth->checkAccess("ADMIN-USERS")) {
         if (isset($args['id'])) {
             $FN = $data['fn'] ?? null;
             $SN = $data['sn'] ?? null;
@@ -40,11 +40,11 @@ $app->patch('/user/{id}', function ($request, $response, $args) {
             $UN = $data['un'] ?? null;
             $PW = $data['pw'] ?? null;
             $Groups = $data['groups'] ?? null;
-            if (!$FN && !$SN && !$EM && !$UN && !$PW && !$Groups) {
-                $ib->api->setAPIResponseMessage('Nothing to update');
-            } else {
+            // if (!$FN && !$SN && !$EM && !$UN && !$PW && $Groups) {
+            //     $ib->api->setAPIResponseMessage('Nothing to update');
+            // } else {
                 $ib->auth->updateUser($args['id'],$UN,$PW,$FN,$SN,$EM,$Groups);
-            }
+            // }
         } else {
             $ib->api->setAPIResponse('Error','id missing from request',400);
         }
@@ -57,7 +57,7 @@ $app->patch('/user/{id}', function ($request, $response, $args) {
 
 $app->delete('/user/{id}', function ($request, $response, $args) {
 	$ib = ($request->getAttribute('ib')) ?? new ib();
-    if ($ib->rbac->checkAccess("ADMIN-USERS")) {
+    if ($ib->auth->checkAccess("ADMIN-USERS")) {
         if (isset($args['id'])) {
             $ib->auth->removeUser($args['id']);
         } else {