+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.6.8
 - Add additional style customisation options
+- Add LDAP Support
 
 ## v0.6.7
 - Cleanup IB tools into new plugin